A lot of Changes
I shifted the data by 1,2 and 3 weeks so that I could predict in advance by that number of weeks. Focused mostly on the 3-week predictions. Created a Neural Network with 0.1101 RMSE which is ~ 11% error. Today was the day of the science fair. I am pushing some commits from the past few days.